reglage de UsersPage

+++ src/main/resources/db/migration/V13_1__create_prices_table.sql
@@ -1,10 +1,10 @@
 CREATE TABLE `prices` (
-    `id` INT(11) NOT NULL AUTO_INCREMENT,
+    `id` BIGINT NOT NULL AUTO_INCREMENT,
     `type` VARCHAR(50) COLLATE utf8mb4_unicode_ci NOT NULL,
     `amount` DOUBLE NOT NULL,
     `start_date` DATETIME NOT NULL,
     `end_date` DATETIME DEFAULT NULL,
-    `representation_id` INT(11) NOT NULL,
+    `representation_id` BIGINT NOT NULL,
     PRIMARY KEY (`id`),
     KEY `fk_price_representation` (`representation_id`),
     CONSTRAINT `fk_price_representation` FOREIGN KEY (`representation_id`) REFERENCES `representations` (`id`) ON DELETE CASCADE
