+++ src/main/resources/db/migration/V9_1__create_artist_type_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `artist_type` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `artist_id` int(11) NOT NULL,
-  `type_id` int(11) NOT NULL,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `artist_id` BIGINT NOT NULL,
+  `type_id` BIGINT NOT NULL,
   PRIMARY KEY(id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
